refactor(logging): convert SessionLogger to functional prefixLogger
Replace the SessionLogger class with a prefixLogger(inner, prefix) factory
that wraps a Logger and prefixes every message, and rename the module to
prefixLogger.ts. The prefix is now generic (a session ID, a workspace name,
etc.); the caller passes the bracketed session prefix. Addresses review
feedback on #1073.
